:sparkles: feat: create useLogin hook for handling login logic and refactor Login component to utilize it
